Moves all these new files into the DataMagic::Index namespace
